fixes for recovery.fstab

+++ tools/device/mkvendor.sh
@@ -60,6 +60,10 @@ then
     cp $BOOTIMAGE $TMPDIR
     pushd $TMPDIR > /dev/null
     unpackbootimg -i $BOOTIMAGEFILE > /dev/null
+    mkdir ramdisk
+    pushd ramdisk > /dev/null
+    gunzip -c ../$BOOTIMAGEFILE-ramdisk.gz | cpio -i
+    popd > /dev/null
     BASE=$(cat $TMPDIR/$BOOTIMAGEFILE-base)
     CMDLINE=$(cat $TMPDIR/$BOOTIMAGEFILE-cmdline)
     PAGESIZE=$(cat $TMPDIR/$BOOTIMAGEFILE-pagesize)
@@ -87,7 +91,10 @@ fi
 for file in $(find $TEMPLATE_DIR -name '*.template')
 do
     OUTPUT_FILE=$DEVICE_DIR/$(basename $(echo $file | sed s/\\.template//g))
-    cat $file | sed s/__DEVICE__/$DEVICE/g | sed s/__MANUFACTURER__/$MANUFACTURER/g | sed -f $TMPDIR/sedcommand | sed s/__BASE__/$BASE/g | sed s/__PAGE_SIZE__/$PAGESIZE/g > $OUTPUT_FILE
+    if [ ! -f $OUTPUT_FILE ]
+    then
+        cat $file | sed s/__DEVICE__/$DEVICE/g | sed s/__MANUFACTURER__/$MANUFACTURER/g | sed -f $TMPDIR/sedcommand | sed s/__BASE__/$BASE/g | sed s/__PAGE_SIZE__/$PAGESIZE/g > $OUTPUT_FILE
+    fi
 done
 
 mv $DEVICE_DIR/device.mk $DEVICE_DIR/device_$DEVICE.mk
